+++ Cargo.toml
@@ -9,6 +9,9 @@ repository = "https://github.com/sunng87/rustmann"
 readme = "README.md"
 edition = "2018"
 
+[features]
+tls = ["tokio-rustls", "webpki"]
+
 [dependencies]
 tokio = "=0.2.0-alpha.2"
 futures-core-preview = "=0.3.0-alpha.18"
@@ -16,7 +19,10 @@ futures-util-preview = "=0.3.0-alpha.18"
 protobuf = { version = "2", features = ["with-bytes"] }
 bytes = "0.4"
 derive_builder = "0.7.0"
+getset = "0.0.7"
 failure = "0.1"
+tokio-rustls = { version = "0.12.0-alpha.1" }
+webpki = {version = "0.21" }
 
 [build-dependencies]
 protoc-rust = "2"
