+++ examples/interactive.rs
@@ -7,7 +7,7 @@ use tokio::sync::mpsc;
 
 use protobuf::Chars;
 use rustmann::protos::riemann::Event;
-use rustmann::{RiemannClient, RiemannClientOptions, RiemannClientError};
+use rustmann::{RiemannClient, RiemannClientError, RiemannClientOptions};
 
 #[tokio::main]
 async fn main() -> Result<(), RiemannClientError> {
