+++ src/client.rs
@@ -1,19 +1,17 @@
 use std::future::Future;
 use std::io;
-use std::net::SocketAddr;
 use std::ops::DerefMut;
 use std::pin::Pin;
-use std::str::FromStr;
 use std::sync::{Arc, Mutex};
 use std::task::{Context, Poll};
 
-use derive_builder::Builder;
 use futures_core::future::BoxFuture;
 use futures_util::FutureExt;
 use protobuf::Chars;
 
 use crate::connection::Connection;
 use crate::error::RiemannClientError;
+use crate::options::RiemannClientOptions;
 use crate::protos::riemann::{Event, Query};
 
 #[derive(Clone)]
@@ -58,9 +56,11 @@ impl Future for Inner {
                 }
             },
             ClientState::Disconnected => {
-                let mut f =
-                    Connection::connect(self.options.address, self.options.connect_timeout_ms)
-                        .boxed();
+                let mut f = Connection::connect(
+                    self.options.address().clone(),
+                    *self.options.connect_timeout_ms(),
+                )
+                .boxed();
                 if let Poll::Ready(Ok(conn)) = f.poll_unpin(cx) {
                     let conn_wrapper = Arc::new(Mutex::new(conn));
                     *inner_state = ClientState::Connected(conn_wrapper.clone());
@@ -70,24 +70,6 @@ impl Future for Inner {
                     Poll::Pending
                 }
             }
-        }
-    }
-}
-
-#[derive(Debug, Builder, Clone, Copy)]
-#[builder(setter(into))]
-pub struct RiemannClientOptions {
-    address: SocketAddr,
-    connect_timeout_ms: u64,
-    socket_timeout_ms: u64,
-}
-
-impl Default for RiemannClientOptions {
-    fn default() -> RiemannClientOptions {
-        Self {
-            address: SocketAddr::from_str("127.0.0.1:5555").unwrap(),
-            connect_timeout_ms: 2000,
-            socket_timeout_ms: 3000,
         }
     }
 }
@@ -105,7 +87,7 @@ impl RiemannClient {
 
     /// Send events to riemann via this client.
     pub async fn send_events(&mut self, events: Vec<Event>) -> Result<bool, RiemannClientError> {
-        let timeout = self.inner.options.socket_timeout_ms;
+        let timeout = *self.inner.options.socket_timeout_ms();
         let state = self.inner.state.clone();
         let inner = &mut self.inner;
 
@@ -123,8 +105,11 @@ impl RiemannClient {
     }
 
     /// Query riemann server by riemann query syntax via this client.
-    pub async fn send_query(&mut self, query_string: &str) -> Result<Vec<Event>, RiemannClientError> {
-        let timeout = self.inner.options.socket_timeout_ms;
+    pub async fn send_query(
+        &mut self,
+        query_string: &str,
+    ) -> Result<Vec<Event>, RiemannClientError> {
+        let timeout = *self.inner.options.socket_timeout_ms();
         let state = self.inner.state.clone();
         let inner = &mut self.inner;
