+++ src/lib.rs
@@ -5,7 +5,9 @@ mod client;
 mod codec;
 mod connection;
 mod error;
+mod options;
 pub mod protos;
 
-pub use crate::client::{RiemannClient, RiemannClientOptions, RiemannClientOptionsBuilder};
+pub use crate::client::RiemannClient;
 pub use crate::error::RiemannClientError;
+pub use crate::options::{RiemannClientOptions, RiemannClientOptionsBuilder};
